+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Better test coverage, including testing on Windows.
+- Warn when a newer version exists.
+
 ## [v0.1.2](https://github.com/naacl2022-reproducibility-track/naacl-utils/releases/tag/v0.1.2) - 2021-12-08
 
 ### Fixed